+++ .env.example
@@ -129,11 +129,6 @@ export BOOKWAREHOUSE_NAMESPACE=bookwarehouse
 # Default: debug
 # export ENVOY_LOG_LEVEL=debug
 
-# optional: ROUTES_V2 (true/false) - Enables experimental routes v2 feature
-# Default: false
-# export ROUTES_V2=false
-#--------------------------------------
-
 ### The section below configures certificates management
 ### OSM has 2 ways to manage certificates
 ### Set CERT_MANAGER to "tresor" to use the internal system (relies on k8s secrets)
