+++ charts/osm/templates/osm-deployment.yaml
@@ -48,9 +48,6 @@ spec:
             "--cert-manager-issuer-name", "{{.Values.OpenServiceMesh.certmanager.issuerName}}",
             "--cert-manager-issuer-kind", "{{.Values.OpenServiceMesh.certmanager.issuerKind}}",
             "--cert-manager-issuer-group", "{{.Values.OpenServiceMesh.certmanager.issuerGroup}}",
-            {{- if .Values.OpenServiceMesh.enableRoutesV2Experimental }}
-            "--enable-routes-v2-experimental",
-            {{- end }}
           ]
           resources:
             limits:
