+++ charts/osm/values.yaml
@@ -61,8 +61,6 @@ OpenServiceMesh:
   enableDebugServer: false
   # -- Enable permissive traffic policy mode
   enablePermissiveTrafficPolicy: false
-   # -- Enable experimental routes feature
-  enableRoutesV2Experimental: false
   # -- Enable egress in the mesh
   enableEgress: false
   # -- Deploy Prometheus
