+++ cmd/cli/install_test.go
@@ -831,7 +831,6 @@ func getDefaultInstallCmd(writer io.Writer) installCmd {
 		enableEgress:                  defaultEnableEgress,
 		enablePermissiveTrafficPolicy: defaultEnablePermissiveTrafficPolicy,
 		clientSet:                     fake.NewSimpleClientset(),
-		enableRoutesV2Experimental:    defaultEnableRoutesV2Experimental,
 		deployPrometheus:              defaultDeployPrometheus,
 		enablePrometheusScraping:      defaultEnablePrometheusScraping,
 		deployGrafana:                 defaultDeployGrafana,
@@ -871,7 +870,6 @@ func getDefaultValues() map[string]interface{} {
 				}},
 			"enableDebugServer":             defaultEnableDebugServer,
 			"enablePermissiveTrafficPolicy": defaultEnablePermissiveTrafficPolicy,
-			"enableRoutesV2Experimental":    defaultEnableRoutesV2Experimental,
 			"enableEgress":                  defaultEnableEgress,
 			"deployPrometheus":              defaultDeployPrometheus,
 			"enablePrometheusScraping":      defaultEnablePrometheusScraping,
