+++ pkg/featureflags/featureflags.go
@@ -6,9 +6,6 @@ import (
 
 // OptionalFeatures is a struct to enable/disable optional features
 type OptionalFeatures struct {
-	// RoutesV2
-	RoutesV2 bool
-
 	WASMStats bool
 }
 
@@ -32,11 +29,6 @@ func IsFeatureNameEnabled() bool {
 	return Features.FeatureName
 }
 */
-
-// IsRoutesV2Enabled returns a boolean indicating if the experimental routes feature is enabled
-func IsRoutesV2Enabled() bool {
-	return Features.RoutesV2
-}
 
 // IsWASMStatsEnabled returns a boolean indicating if custom stats will be generated by a WASM extension to Envoy
 func IsWASMStatsEnabled() bool {
